Reorganize the unit test structure
